👌 CLI: Give feedback for `configure-rabbitmq`
Currently the `verdi profile configure-rabbitmq` command doesn't give any feedback to
the user whether the provided options can successfully connect to the RabbitMQ server.
Here we adapt the `detect_rabbitmq_config` function to accept the broker configuration
as `**kwargs`, and use it to check if the provided options in the `configure-rabbitmq`
can successfully connect to the RabbitMQ server. A "success" message is printed if
we can connect to the server, else a warning is printed and the user is asked for
confirmation before proceeding to configure the broker.

A `--force` flag is also added to avoid asking for confirmation in case the command
is unable to connect to the broker.