Engine: Catch `NotImplementedError`in `get_process_state_change_timestamp`
The `get_process_state_change_timestamp` utility calls the method
`get_global_variable` on the storage backend to get the timestamp of the
latest process state change, which is typically stored in the
`db_dbsetting` table. However, not all storage plugins implement, most
notably the `core.sqlite_zip` plugin. Since this is read-only, the
settings table is never used and requesting the timestamp of the last
process state change does not make sense.

Since this utility is used in `verdi process list`, the command would
error since the `NotImplementedError` was not caught. This is now the
case and `verdi process list` will show "never" as the last state change.