Tests: Fix `tests.orm.nodes.test_node:test_delete_through_backend` (#6496)
The test was failing with a `core.sqlite_dos` storage plugin for the
test profile. The problem is that the last assert was checking that the
logs for `data_two` were deleted because `data_two` itself had been
deleted. However, since it was deleted, the ORM instance can no longer
be used either, which was causing an exception. Instead, its pk should
be recorded before deleting the node, and the final check should just
use the pk directly.

It is not quite clear why this test was not failing for the default
`core.psql_dos` storage plugin that is used for tests. It should not be
backend specific since both use SQLAlchemy for the ORM.