Dependencies: Update requirement to `psycopg~=3.0` (#6362)
The `psycopg` library is used by `sqlalchemy` to connect to the
PostgreSQL server. So far, the `psycopg2` package was used, but the
next generation v3 has already been out for a while. The release comes
with a number of performance improvements according to the author.
Although there is no clear timeline for support of v2, we are not
waiting and switching now to the new version.

When v2 was released, instead of releasing a new major version, it was
released as a new library changing the name from `psycopg` to `psycopg2`
but for v3 they are switching back to `psycopg`. This means that users
would still be able to run v2 and v3 along side one another in a single
Python environment. This supports the decision to move to v3.

Note that `aiida-core` will not be supporting both at the same time and
from now only supports v3. Interestingly, from the very early versions
of AiiDA, the profile would contain the `database_engine` key. This is
still an option in `verdi setup` and `verdi quicksetup`, except it is
not really an option as it is hardcoded to `postgresql_psycopg2`.
So all `core.psql_dos` profiles out there contain:

    'main': {
        'storage': {
            'backend': 'core.psql_dos',
            'config': {
                'database_engine': 'postgresql_psycopg2',
                ...
            }
        },
    }

The value is not actually used however as the connection string for
sqlalchemy's engine, where it _could_ be used, simply hardcodes this to
`postgres://` which is the default psycopg dialect and maps to
`postgres+psycopg2://`. This value is now simply updated to
`postgres+psycopg://` to target the new version of `psycopg`.

Because it is hardcoded, a migration for the existing configs is not
required and it can be left as a vestigial attribute. Since `verdi setup`
and `verdi quicksetup` are deprecated now anyway, the options don't have
to be removed here.