Dependencies: Pin requirement to minor version `sphinx~=7.2.0` (#6527)
The recently released v7.4.0 causes the build of the documentation to
fail with warnings of the type:

    Failed to get a method signature for ...: unhashable type

It also leads to issues with incremental builds where it fails if
a build exists that was generated with an older version.

Both these bugs were fixed in v7.4.1 and v7.4.2 respectively, but more
warnings remain. Instead of waiting for these to be fixed, we are just
pinning to a minor version here which should anyway be done for optional
dev dependencies. There is no need to automatically update those to
latest releases.

Cherry-pick: a610460572f2d85435dfa850fdc247e1c666ef68