CLI: Fix bug `verdi presto` when tab-completing without config
For a clean environment where the config directory had not yet been
created, tab-completing `verdi presto` would result in an exception
being thrown because the callable for the default of the
`--profile-name` option would try to access the configuration.

The exception is now caught and the callable simply returns the default
profile name `presto` which should be correct since no profiles should
exist if there is not even a configuration directory.