Docs: Add succint overview of limitations of no-services profile
A table is added to the quick installation guide that gives a more
succinct overview of the limitations of a profile created without a
broker and with SQLite instead of PostgreSQL. This was already explained
in more detail in text, but that was not complete and could be too dense
for users, making them likely to skip it.

The code is updated to provide a link to this documentation section
whenever an error is displayed that a broker is not configured. This way
users can try to understand why the functionality they are trying to use
is not supported and how, if they really care about it, they can go
about still installing and configuring RabbitMQ after the fact.

Cherry-pick: 4038d550452df25cccd13137b9dfb5823345c544