Docs: Remove note in installation guide regarding Python requirement
This admonition is not that useful since it still refers the user to an
external website and the target information is not immediately obvious
either. Besides, if the installed Python version is not supported, this
will automatically be reported by the package manager.

Cherry-pick: 96d9fbdc0027b3fd357a6aab172ec6f7f3cc94af