Engine: Ignore failing process state change for `core.sqlite_dos`
For each process state change, the engine calls the utility function
`aiida.engine.utils.set_process_state_change_timestamp`. This calls
`set_global_variable` on the storage plugin to update the
`process|state_change|.*` key in the settings table. This value is used
in `verdi process list` to show when the last time a process changed its
state, which serves as a proxy of daemon activity.

When multiple processes would be running, this call would throw an
exception for the `core.sqlite_dos` storage plugin. This is because
SQLite does not support concurrent writes that touch the same page,
which is the case when multiple writes are updating the same row.

Since the updating of the timestamp is not crucial for AiiDA functioning
properly, especially since it is because another process was trying to
perform the same update, it is safe to ignore the failed update and
simply log that as a warning.