Engine: Change signature of `set_process_state_change_timestamp`
The function took a `Process` instance but then only uses it to fetch
its associated node. Since a `Process` instance is way more difficult to
mock, it makes testing the function unnecessarily complicated. Since it
only needs the process node, the signature is changed to accept the node
instead of the process. This utility function is unlikely to be used in
client code, justifying this technically backwards incompatible change.