CLI: Validate storage in `verdi storage version`
The `verdi storage version`, in addition to printing the version of the
code's and storage's schema, now also validates the storage. If the
storage is corrupt or cannot be reached, the command returns the exit
code 3. If the storage and code schema versions are incompatible, exit
code 4 is returned. This way this command serves as an alternative to
running `verdi storage migrate` as a way to check whether a profile
needs to be migrated. The `verdi storage migrate` command needs to
perform checks such as whether the daemon is running and so is always
going to be slower.