CLI: Fix `verdi storage migrate` for profile without broker (#6550)
The command needs to make sure the daemon of the profile is not running
so it instantiates the `DaemonClient` but this raises for profiles that
do not define a broker. Since the daemon cannot be started for
brokerless profiles anyway the command does not have to check in this
case.

Cherry-pick: 7402c17755b332cc9a06e88516621e575f0d1cce