CLI: Fix exception for `verdi plugin list`
In e952d7717c1d8001555e8d19f54f4fa349da6c6e a bug in `verdi plugin list`
was fixed where the conditional to check whether the plugin was a
process class would always raise an `AttributeError` if the plugin was
not a `Process` or a proces function. As a result, the code would never
get to the else-clause.

The else-clause contained itself another bug, which was now revealed by
the fixing of the bug in the conditional. The else-clause would call the
`get_description` classmethod of the plugin, but no classes in AiiDA
that are pluginnable even define such a class method. Probably, the
original author confused it with the instance method `get_description`
but the `verdi plugin list` command just deals with the class.

The `get_description` call is replaced with just getting `__doc__` which
returns the docstring of the class/function, or `None` if it is not
defined. In the latter case, a default message is displayed saying that
no description is available.

Since the else-clause code was never reached before the recent fix and
the `get_description` method was never supported officially by AiiDA's
pluginnable interfaces, it is fine to just change this behavior.