Tests now use ssh key specified by env var AIIDA_PYTEST_SSH_KEY
The tests were relying on `~/.ssh/id_rsa` to be used to allow ssh to
localhost. This had the side effect that several tests specified a ssh
key in their configuration that was then not working in the
authentication but a fall back to the default ssh key still allowed a
connection. This PR fixes this behaviour and uses also the ssh key
specified for the connection. This also allows the user to specify a
dedicated ssh key for aiida tests and does not enforce the usage of the
default key.