Add `property` logic for `filepath_files`
of `PortableCode`.

Some notes on the implementation:
- Removed the `put_object_from_tree` from the `__init__`, as this is
  done in the `setter` which is anyway called when the `property` is
  actually set in the `__init__`.
- Added check for absolute Path (would otherwise be
captured in `Repository.put_object_from_tree`, but maybe better to catch
here already).
- Not sure if the repository should be erased before putting
another `object_from_tree`. The logic I followed is that if one wants to
set another `filepath_files` to not keep the previous ones and get a
polluted repository, but rather just setting the new one. However, it
should be avoided ofc that unwanted things are removed from the
repository. And, if one wants to add multiple directories, etc., this
situation is described in the constructor docstring, and one can use the
methods of the `NodeRepository`, so I guess this shouldn't be done via
setting the `filepath_files`