WIP: Implement force-kill option in verdi process kill
This allows plumpy ignore the callback so it does not get stuck in it
but directly transitions the process to killed. The disadvantage is that
the job behind the process might not be killed as we do not wait for.
But I think that can be clearly communicated to the user in the help
message and it has a similar logic as `kill -9 <PID>`.

TODO: As you mentioned @kh with this solution the worker process will
not be killed. Some parts of plumpy in stepping are needed to kill
the process, but I am not sure which ones.