Set `extra_files` for PortableCode export
This is done rather than calling `NodeRepository.copy_tree()` inside the
`_prepare_yaml` function to adhere to the standard that the `_prepare_`
methods should return a tuple with the contents of the main export file
as the first element, and a dictionary with the output file paths as
strings as keys and their utf-8 encoded contents as values. The creation
of these `extra_files` is then handled in the `export` method of
`orm.Data`.

We create the necessary parent directories for the `extra_files` here in
the overwritten `_prepare_yaml` method of `orm.PortableCode` as to not
modify the original `export` method that applies to all `orm.Data`
nodes.