Avoid duplicate dumping of calculations
which were called by a workflow that is part of a group. Without
extending the `nodes_in_groups` by the `called_descendants` of the
`WorkflowNode`s, a top-level directory `calculations` will be created
and the called `CalculationNode`s of a workflow that is assigned a group
would be dumped. This leads to duplication, as the calculations are
already being dumped in the respective `workflows` directory within the
group directory.