DevOps: Fix json query in reading the docker names to filter out fields not starting with aiida (#6573)
With the version update of bake-action the `BAKE_METADATA` a field with warning
information (buildx.build.warnings) was added that does not contain the key
`image.name` so the json query failed. With this commit another json query was
added that filters out every field name that does not start with "aiida", so
the field with warning information is filtered out.

Co-authored-by: Jusong Yu <jusong.yeu@gmail.com>

Cherry-pick: e1467edca902867e53605e0e60b67f8767bf8d3e