CLI: Handle `None` process states in build_call_graph
The process definition allows a None state but the build_call_graph
utils function intended for a creating printable output of the call
graph does crash in this case.  A None state indicates some undefined
behavior in the code, but the application should not fail because of the
formatting of the output.  This commit makes build_call_graph handle
None states. Partial fix for #6585.