Dump only `sealed` by default and add `incremental` option
Previously, no checks were applied on the state of a process. Now, by
default, only `sealed` processes are being dumped to disk. This can be
disabled with the new `--dump-unsealed` flag. In addition, the option of
`--incremental` dumping is added, in which case an incomplete existing
dumping output directory can be gradually filled up with data, e.g.,
while the process is running in conjunction with the `--dump-unsealed`
flag. Before, only the `--overwrite` flag was available for this
purpose, which would directly clean the entire top-level directory.

Lastly, the `validate_make_dump_path` method was refactored out of the
`ProcessDumper` and put as a normal function into `utils.py` in the
dumping directory.