`DirectScheduler`: Ensure killing child processes (#6572)
The current implementation only issues a kill command for the
parent process, but this can leave child processes orphaned. The
child processes are now retrieved and added explicitly to the
kill command.

Cherry-pick: fddffca67b4f7e3b76b19df7db8e1511c449d2d9

Edits: Downgrades scheduler usage to old API in fix #6572

The fix in #6572 was pushed after the `Scheduler` API has been
refactored in PR #6043. To not include this breaking change in a minor
release we adapt the usage of `Scheduler` in the PR to the old API.