CLI: Dump only `sealed` process nodes (#6591)
Previously, no checks were applied on the state of a process when dumping its data to disk via `verdi process dump`.
Now, by default, only `sealed` processes are being dumped. This can be disabled with the new `--dump-unsealed` flag. In
addition, the option of `--incremental` dumping is added, in which case an incomplete existing dumping output directory
can be gradually filled up with data, e.g., while the process is running in conjunction with the `--dump-unsealed` flag.
Before, only the `--overwrite` flag was available to handle the case of an already existing directory, however, it
directly cleans the entire top-level directory. Lastly, the `prepare_dump_path` method (previously
`validate_make_dump_path`) was refactored out of the `ProcessDumper` and put as a normal function into `utils.py` in the
`tools/dumping` directory.