Scheduler: Allow a memory specification of zero for the SLURM plugin (#6605)
In the SLURM memory specification, a value of zero is treated as a special
case indicating that no memory limit should be used.
This is documented here: https://slurm.schedmd.com/sbatch.html under the
'--mem' section.

However, currently the SLURM plugin demands a positive integer, but this logic
puts an unexpected limitation on the SLURM plugin. This commit changes
this logic to allow a value of 0 to be accepted.