CLI: Check user execute in `verdi code test` for `InstalledCode` (#6597)
An additional test is added for user execute permissions in the `validate_filepath_executable` method of `InstalledCode` called by `verdi code test`.

This can help debugging (or even avoiding) the issue of `touch`ing a dummy code, registering it in AiiDA, and then not being able to use it due to the script not being executable (as default Linux file permissions are `rw-rw-r--`). Some of us have actually encountered this while developing, and the resulting issue was difficult to debug. With this change, one should be able to find it quicker using `verdi code test`.