Mark private API with an underscore
The AiiDA public API has been defined in the documentation as every import that
can be done from a secondary level. This is however not transparent to users
and plugin developers as it is hidden in the doc resulting in a lot of misuse
of private functionalities. The widespread use of these private functionalities
blocks improvements in the code as we are more hesitant in changing
functionalities for wrong backwards compatibility.

This PR marks clearly the internal functioanlities with an underscore so a
misuse only can happen intentionally. This gives us the opportunity to improve
the public API as we can implement replacements for functionalities in the
private API that are widely used.