Typing: mypy fix for `orm.List` the `pop` and `index` methods (#6635)
The function signature of `pop` and `index` methods of `orm.List` is updated to synchronous with `MutableSequence` which it inherit from. The implementation of `orm.List` is expected to be the same as vanilla python `list` type. 
The new function signature has the index parameter for `pop` with default `-1` for pop the last element from the list. 

This change might be a break change because the previous implementation can be used differently.