CLI: Handle `None` process states in build_call_graph (#6590)
The process definition allows a None state but the `build_call_graph`
utils function intended for creating a printable output of the call
graph does crash in this case. A None state indicates an undefined
behavior in the code, but the application should not fail because of the
formatting of the output. This commit makes `build_call_graph` handle
`None` states. Partial fix for #6585.