Replace `remote_data` fixtures by factory
This allows just defining one fixture and passing a mode parameter, to
create different `RemoteData` instances, one with SSH and one without.
In addition, it allows to pass different content directly to the fixture
factory to create differently parametrized instances. Without a factory,
this is not possible, as the instantiated `RemoteData` that is returned
cannot be parametrized (by something like `(content=b'a')`), as it would
be already the instantiated `RemoteData` object.

In addition, this change also removes the need for the
`request.getfixturevalue(fixture)` code, but allows just passing the
fixture factory and the mode as a parameter.