ORM: Add `get_size_on_disk` method to `RemoteData` (#6584)
By default, the new `get_size_on_disk` method of `RemoteData` calls the private `_get_size_on_disk_du` that uses `du` to obtain the total directory size in bytes. If the call to `du` fails for whatever reason, recursive `stat` is being used via `_get_size_on_disk_stat`. This route somewhat discouraged, and a warning is issued, as `stat` returns the apparent size of files, not the actual disk usage.

In addition, the CLI endpoint `verdi data core.remote size` is added, and the tests for `RemoteData` are expanded and parametrized to use both, `LocalTransport` and `SshTransport`, testing the functionality added in this PR.