Update uv.lock + CI tweaks + fix pymatgen issue (#6676)
Recreate the uv lockfile (`uv.lock`) with a newer uv version that uses an updated resolution strategy.

Previously, it tried to find a version of the package that would be compatible with all supported Python versions, which in turn meant that if a package dropped support for 3.9 for example, we'd be stuck on an old version.
In the new uv version, this is not the case, and uv will try to find the newest version of a package for a given python version.

Incidentally, this also help partly resolve the pymatgen issue, see #6680. The issue has been fixed in a new version of monty package, which however does not support Python 3.9. So in the updated lockfile, the issue is fixed for Python >=3.10.

For Python 3.9, a workaround suggested in materialsproject/pymatgen#4243, and unset the CI environmental variable.
Since this workaround is quite targeted (only in two specific job steps, and only for Python 3.9), I believe this is acceptable, and we will be dropping Python 3.9 support sometimes this year anyway.