Use uv lockfile in readthedocs build (#6685)
Now that we have a bit of good experience with uv lockfile, it would be nice to use it in more places to make various CI jobs more stable. Here I am using it for the RTD docs build. We've already been using uv there so that change is relatively straightforward.

Coincidentaly, ReadTheDocs made some updates to make the configuration with uv nicer, just a couple of days ago see readthedocs/readthedocs.org#11289 (comment) for details.

This also updates the uv version in RTD config.