Adapt message arguments passing to process controller (#6668)
Changes required after aiidateam/plumpy#301 and aiidateam/plumpy#291

For the rpc call of process.kill and process.pause, the passed in argument name is changed from msg to msg_text. Same for kill_processes and pause_processes of process controller.
Meanwhile, when calling _perform_actions the action is passed as a callable that takes no argument, it then requires the callable should be constructed with its arguments using functools.partial before hands. This increase the readability by makes the arguments stay together with the function itself where originally the arguments passed to callable is separated from other arguments for _perform_action itself.

- bump plumpy version to `0.24.0`.