CI: fix failed test_containerized.py integration test for containerized code (#6707)
fixes #6709

The test is run in nightly.yml as integration test for running containerized code.
The problem may because the new ubuntu don't allow the container to get root permission even for singularity. Since singularity is anyway outdated and we support docker as containerized code engine. I change it to use docker to run the code from a image.
This also remove the CI action (which has no maintains for around 2 years) of building singularity.