Set minimum uv version in pyproject.toml (#6714)
Closes #6712 

This sets the minimum uv version pyproject.toml, which is needed to have a consistent lockfile. 
When a dev has an outdated uv version, they'll get an error

```
error: Required version `>=0.5.20` does not match the running version `0.5.19`
```

Also pinned the `setup-uv` action to a specific version to prevent accidental regressions since that action is still undergoing heavy development. The new 5.2.0 also allows us to drop a VIRTUAL_ENV workaround. :tada: