Dependencies: version check on `sqlite` C-language (#6567)
Enforce the minimum supported version of `sqlite` in `C`-language.
Check the version of sqlite3 (in C) wherever relevant. Mainly when, creating or initializing a new profile with either `sqlite_zip` or `sqlite_dos` backend.

Therefore it also appears during `verdi status` (only if the default profile is with `sqlite` storage backend)

We decided that the check should not happen during AiiDA installation, matching `sqlite` versions could be painful, and we want to keep installation as smooth as possible. Moreover, not everyone needs to be bothered, as not everyone uses `sqlite` backend.