Fix verdi devel check-undesired-imports when tui extra is installed (#6693)
`verdi devel check-undesired-imports` makes sure that we don't have heavy imports during verdi startup to keep the CLI snappy (especially for tab completion).

One of the expensive modules that it checks is `asyncio`. Unfortunately, when you install the tui extra, the trogon package that powers the tui functionality seems to use asyncio, and this makes the test fail.
(indeed, one of the TUIs current downsides is that it makes all CLI interactions slower, even if you don't use the TUI subcommand).

The PR makes the test more clever and don't check for `asyncio` import when `tui` extras is installed.