ORM: Use `skip_orm` as the default implementation for `SqlaGroup.add_nodes` and `SqlaGroup.remove_nodes` (#6720)
This commit removes the `skip_orm` flag and adopts the optimized approach behind it as the default behavior for the `add_nodes` and `remove_nodes` methods.

As discussed in #5453, the ORM-based implementations of `add_nodes` and `remove_nodes` were found to be inefficient, and to address this, the `skip_orm` flag was introduced. This flag enabled a faster, core API-based approach for these operations, but was never well documented or advertised.

All existing tests pass with the new implementation, and are sufficiently comprehensive to ensure its safety. Consequently, it is a reasonable decision to make the `skip_orm` implementation the default for bulk insertion and removal of elements within a group.